Migrate STT Engines to new STTPlugin api
