+++ plugins/stt/google-stt/__init__.py
@@ -1,0 +1,1 @@
+from .google import GoogleSTTPlugin
