+++ plugins/stt/julius-stt/plugin.info
@@ -1,0 +1,10 @@
+[Plugin]
+Name = julius-stt
+Version = 1.0.0
+License = MIT
+URL = http://jasperproject.github.io/
+Description = A very basic Speech-to-Text engine using Julius.
+
+[Author]
+Name = Jasper Project
+URL = http://jasperproject.github.io/
