+++ plugins/stt/pocketsphinx-stt/sphinxvocab.py
@@ -2,13 +2,13 @@
 import os
 import logging
 import tempfile
-from g2p import PhonetisaurusG2P
 try:
     import cmuclmtk
-except ImportError:
-    logging.getLogger(__name__).error("Error importing CMUCLMTK module. " +
-                                      "PocketsphinxVocabulary will not work " +
-                                      "correctly.", exc_info=True)
+except:
+    pass
+
+from client import jasperpath
+from .g2p import PhonetisaurusG2P
 
 
 def get_languagemodel_path(path):
@@ -27,7 +27,7 @@ def get_dictionary_path(path):
     return os.path.join(path, 'dictionary')
 
 
-def compile_vocabulary(directory, phrases, config):
+def compile_vocabulary(config, directory, phrases):
     """
     Compiles the vocabulary to the Pocketsphinx format by creating a
     languagemodel and a dictionary.
@@ -38,7 +38,24 @@ def compile_vocabulary(directory, phrases, config):
     logger = logging.getLogger(__name__)
     languagemodel_path = get_languagemodel_path(directory)
     dictionary_path = get_dictionary_path(directory)
-    g2pconverter = PhonetisaurusG2P(config)
+
+    try:
+        executable = config['pocketsphinx']['phonetisaurus_executable']
+    except KeyError:
+        executable = 'phonetisaurus-g2p'
+
+    try:
+        nbest = config['pocketsphinx']['nbest']
+    except KeyError:
+        nbest = 3
+
+    try:
+        fst_model = config['pocketsphinx']['fst_model']
+    except KeyError:
+        fst_model = os.path.join(jasperpath.APP_PATH, os.pardir,
+                                 'phonetisaurus', 'g014b2b.fst')
+
+    g2pconverter = PhonetisaurusG2P(executable, fst_model, nbest)
 
     logger.debug('Languagemodel path: %s' % languagemodel_path)
     logger.debug('Dictionary path:    %s' % dictionary_path)
@@ -71,6 +88,7 @@ def compile_languagemodel(text, output_file):
 
     # Create language model from text
     logger.debug("Creating languagemodel file: '%s'", output_file)
+    print("TEXT", text)
     cmuclmtk.text2lm(text, output_file, vocab_file=vocab_file)
 
     # Get words from vocab file
