+++ plugins/stt/witai-stt/plugin.info
@@ -1,0 +1,10 @@
+[Plugin]
+Name = witai-stt
+Version = 1.0.0
+License = MIT
+URL = http://jasperproject.github.io/
+Description = Speech-To-Text implementation which relies on the Wit.ai Speech API.
+
+[Author]
+Name = Jasper Project
+URL = http://jasperproject.github.io/
