+++ plugins/stt/witai-stt/witai.py
@@ -1,0 +1,67 @@
+import logging
+import requests
+from client import plugin
+
+
+class WitAiSTTPlugin(plugin.STTPlugin):
+    """
+    Speech-To-Text implementation which relies on the Wit.ai Speech API.
+
+    This implementation requires an Wit.ai Access Token to be present in
+    profile.yml. Please sign up at https://wit.ai and copy your instance
+    token, which can be found under Settings in the Wit console to your
+    profile.yml:
+        ...
+        stt_engine: witai
+        witai-stt:
+          access_token:    ERJKGE86SOMERANDOMTOKEN23471AB
+    """
+
+    def __init__(self, *args, **kwargs):
+        plugin.STTPlugin(self, *args, **kwargs)
+        self._logger = logging.getLogger(__name__)
+        self.token = self.profile['witai-stt']['access_token']
+
+    @property
+    def token(self):
+        return self._token
+
+    @token.setter
+    def token(self, value):
+        self._token = value
+        self._headers = {'Authorization': 'Bearer %s' % self.token,
+                         'accept': 'application/json',
+                         'Content-Type': 'audio/wav'}
+
+    @property
+    def headers(self):
+        return self._headers
+
+    def transcribe(self, fp):
+        data = fp.read()
+        r = requests.post('https://api.wit.ai/speech?v=20150101',
+                          data=data,
+                          headers=self.headers)
+        try:
+            r.raise_for_status()
+            text = r.json()['_text']
+        except requests.exceptions.HTTPError:
+            self._logger.critical('Request failed with response: %r',
+                                  r.text,
+                                  exc_info=True)
+            return []
+        except requests.exceptions.RequestException:
+            self._logger.critical('Request failed.', exc_info=True)
+            return []
+        except ValueError as e:
+            self._logger.critical('Cannot parse response: %s',
+                                  e.args[0])
+            return []
+        except KeyError:
+            self._logger.critical('Cannot parse response.',
+                                  exc_info=True)
+            return []
+        else:
+            transcribed = [text.upper()]
+            self._logger.info('Transcribed: %r', transcribed)
+            return transcribed
